Fix the previous/next photo test
First, explicit IDs in the fixture are needed, because the ordering
is based on these IDs and if we rely on automatic ID assignment, our
test may fail or pass randomly.

Second, better use the exact photos from the fixure, instead of
duplicating the prev/next loading via the album logic in the test.